# docsForAI/StackOpenAPI.yml
openapi: 3.0.0
info:
  title: Disk API
  description: |
    The RESTful API for the webinterface and apps to connect and interact with users and files on the disk.

    ### Disclaimer

    We provide the APIs as-is. That means we don’t provide warranties of any kind, either express or implied.
    You can use this API but it may change at any point in time without advance notice.

    ### Headers

    **Note on header names**: Throughout this API specification headers are considered case-insensitive as
    [RFC2616](https://www.ietf.org/rfc/rfc2616.txt) dictates.
    This means that header name casing may not  be exactly as you will find them documented.
    This is due to Go's header set code which normalizes them so only the first letter and each letter after a dash (-)
    is capitalized. For instance the `X-Pre2FAToken` will actually be sent as `X-Pre2fatoken`. For improved readability
    we have documented the former however.

    ### Rate limits

    First of all, the rate limits are meant as a measure of keeping the service available to everyone, so any disk deemed
    to be using too much resources and having negative effects on other disks can be limited further than the limits stated here
    without prior notice.

    #### Request rate limit
    In general there is a limit on the amount of requests done to a disk. This covers all http and sftp requests.
...
    If too many requests are send for a disk, the API will return status code `429 Too Many Requests`.

...

...

...
paths:
...
  /authenticate:
    post:
      summary: User login
      description: >-
        Authenticates a user for the disk and returns a session or pre 2fa
        session token.
      operationId: authenticate
      tags:
      - authentication
      parameters:
      - in: header
        name: X-SetCookie
        description: |
          Set to `true` to let the server return a `SetCookie` header for the newly created token.
          Does not do anything when 2FA is enabled and another login step is required.
        schema:
          type: boolean
      responses:
        '200':
          description: Returned if the user was successfully authenticated.
          headers:
            X-CSRF-Token:
              description: >-
                The CSRF token associated with the session. Required for session
                requests which use a cookie instead of a header.
              schema:
                type: string
            X-2FA-Required:
              description: >-
                True when a 2 factor authentication challenge is required to
                complete the authentication.
              schema:
                type: boolean
            X-2FA-Verify-URL:
              description: >-
                Only returned if 2FA is enabled for this user. The url to complete the login process
                by verification of a 2FA challenge (/api/v2/authenticate/verify).
              example: /api/v2/authenticate/verify
              schema:
                type: string
            X-Pre2FAToken:
              description: >-
                The token used to complete the 2 factor authentication
                challenge. Only set if X-2FA-Required is True.
              schema:
                type: string
            X-SessionToken:
              description: >-
                The token used to authenticate for the rest of the API. Only set
                if X-2FA-Required is False.
              schema:
                type: string
            X-TokenID:
              description: >-
                The database identifier for the created session token. Use it to
                delete the token with the DELETE request.
              schema:
                type: integer
                format: int64
            Set-Cookie:
              schema:
                type: string
              description: An updated session cookie with name `SessionToken`. Only set when the `X-SetCookie` header was set _and_ 2FA is off.
        '400':
          description: Returned if a parameter was invalid.
        '401':
          description: Returned if the login was unsuccessful.
        '429':
          description: Returned if there are too many login attempts for this disk.
        '500':
          description: Returned if the server errored and could not complete the request for unknown reasons.
        '503':
          description: Returned if the disk is currently unavailable. Retry the request later.
        '507':
          description: Too many auth tokens exist for this disk, try removing some.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/authentication'
        description: User credentials and optional session duration in seconds.
        required: true
...
  /disk-events/share:
    get:
      summary: List share disk events
      operationId: listShareEvents
      tags:
        - event
      security:
        - sessionToken: [ ]
        - sessionCookie: [ ]
      description: Filter and list share disk events.
      parameters:
        - in: query
          name: limit
          description: The maximum amount of events to retrieve.
          schema:
            type: integer
            default: 200
            maximum: 1000
        - in: query
          name: fromEventId
          description: The eventId at which to start listing events. Mutually exclusive with `start`.
          schema:
            type: integer
            format: uint64
        - in: query
          name: start
          description: Unix timestamp indicating the start time to get events for. Mutually exclusive with `fromEventId`.
          schema:
            type: integer
            format: int64
        - in: query
          name: end
          description: Unix timestamp indicating the end time to get events for.
          schema:
            type: integer
            format: int64
        - in: query
          name: userId
          description: The user to fetch events for.
          schema:
            type: integer
            format: int64
        - in: query
          name: shareNodeId
          description: The node id shared to fetch events for.
          schema:
            type: integer
            format: int64
        - in: query
          name: eventType
          description: The specific event type to retrieve.
          schema:
            type: string
            enum:
              - create
              - update
              - delete
      responses:
        '200':
          description: Returns a list of events matching given criteria.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/listShareEvents'
        '400':
          description: Bad input parameter.
        '401':
          description: Returned if the user is not logged in.
        '403':
          description: Returned if the user is not allowed to view disk events.
        '429':
          description: Returned if there are too many requests for this disk.
        '500':
          description: Returned if the server errored and could not complete the request for unknown reasons.
        '503':
          description: Returned if the disk is currently unavailable. Retry the request later.
...
  /me/events/share:
    get:
      summary: List my share disk events
      operationId: listMyShareEvents
      tags:
        - event
      security:
        - sessionToken: [ ]
        - sessionCookie: [ ]
      description: Filter and list my share disk events.
      parameters:
        - in: query
          name: limit
          description: The maximum amount of events to retrieve.
          schema:
            type: integer
            default: 200
            maximum: 1000
        - in: query
          name: fromEventId
          description: The eventId at which to start listing events. Mutually exclusive with `start`.
          schema:
            type: integer
            format: uint64
        - in: query
          name: start
          description: Unix timestamp indicating the start time to get events for. Mutually exclusive with `fromEventId`.
          schema:
            type: integer
            format: int64
        - in: query
          name: end
          description: Unix timestamp indicating the end time to get events for.
          schema:
            type: integer
            format: int64
        - in: query
          name: shareNodeId
          description: The node id shared to fetch events for.
          schema:
            type: integer
            format: int64
        - in: query
          name: eventType
          description: The specific event type to retrieve.
          schema:
            type: string
            enum:
              - create
              - update
              - delete
      responses:
        '200':
          description: Returns a list of events matching given criteria.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/listShareEvents'
        '400':
          description: Bad input parameter.
        '401':
          description: Returned if the user is not logged in.
        '429':
          description: Returned if there are too many requests for this disk.
        '500':
          description: Returned if the server errored and could not complete the request for unknown reasons.
        '503':
          description: Returned if the disk is currently unavailable. Retry the request later.
...
  '/share/{shareURLToken}':
    post:
      summary: Authorize share token
      operationId: authorizeShare token
      tags:
      - public-share
      description: >-
        Gets a public share token for interacting with public shares. Optionally requires a password when the share declares it so.
      parameters:
      - in: header
        name: X-SetCookie
        description: Set to `true` to let the server return a `SetCookie` header for the newly created token.
        schema:
          type: boolean
      - in: path
        name: shareURLToken
        description: ID of the share
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/shareToken'
      responses:
        '201':
          description: Returned when the share token was successfully created.
          headers:
            X-ShareToken:
              description: The share token that has been created.
              schema:
                type: string
            X-CSRF-Token:
              description: The CSRF token belonging to this share token.
              schema:
                type: string
        '400':
          description: Returned when the existing share token was not found.
        '401':
          description: Returned when a password was not provided or incorrect.
        '404':
          description: Returned when the share was not found.
        '412':
          description: Returned when the share has expired.
        '429':
          description: Returned if there are too many requests for this disk.
        '500':
          description: Returned if the server errored and could not complete the request for unknown reasons.
        '503':
          description: Returned if the disk is currently unavailable. Retry the request later.
    get:
      summary: Public share info
      operationId: getShareInfo
      tags:
      - public-share
      security:
      - shareToken: []
      - shareSession: []
      description: >-
        Gets a public share information object.
      parameters:
      - in: path
        name: shareURLToken
        description: ID of the share
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Returned if the share was found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getShareByToken'
        '401':
          description: >-
            Returned if the share was found, but the password was not provided
            or incorrect.
        '404':
          description: Returned if the share (token) was not found.
        '429':
          description: Returned if there are too many requests for this disk.
        '500':
          description: Returned if the server errored and could not complete the request for unknown reasons.
        '503':
          description: Returned if the disk is currently unavailable. Retry the request later.
...
  '/share/{shareURLToken}/upload':
    post:
      summary: Public share upload new file
      operationId: UploadSharedFile
      tags:
        - public-share
      security:
        - shareToken: [ ]
        - shareSession: [ ]
      description: |
        The files endpoint provides a way to upload files to your public share.
        It is different from other endpoints as it requires extra data to
        be passed in the header instead of the body as the body is used
        for the to-upload data.
      parameters:
        - in: path
          name: shareURLToken
          description: ID of the share
          required: true
          schema:
            type: string
        - in: header
          name: X-FileByteSize
          required: true
          description: The exact number of bytes of the file you are uploading.
          schema:
            type: integer
            format: int64
            minimum: 0
            maximum: 134217728
        - in: header
          name: X-ParentID
          required: true
          description: The NodeID of the directory to upload this file into.
          schema:
            type: integer
            format: int64
        - in: header
          name: X-Filename
          required: true
          description: |
            The Base64 (RFC 4648) encoded filename for the newly uploaded file. The filename has to not exist yet.
            The filename has a number of restrictions:
              - The maximum length of the characters cannot exceed 148 bytes (for ex. 148 ASCII characters or 37 emojis)
              - The name cannot equal . or ..
              - The name cannot contain unicode control characters
              - The name cannot contain these characters: / \ < > " | ? *
          schema:
            type: string
        - in: header
          name: X-Overwrite
          required: false
          description: >-
            Whether to overwrite any existing file on the given parentID + name.
            Default = false.
          schema:
            type: boolean
        - in: header
          name: X-Modified
          required: false
          description: >-
            Optional unix timestamp containing the file modification date to
            set.
          schema:
            type: integer
            format: int64
        - in: header
          name: X-SHA1-Checksum
          required: false
          description: >-
            Optional SHA1 HEX checksum of the contents of the file to be uploaded. The file contents will be hashed and checked prior to committing the file to the disk.
          schema:
            type: string
      requestBody:
        description: The contents of the file as raw bytes, exactly the amount as specified in `X-FileByteSize`.
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        '201':
          description: Returned if the file was uploaded successful.
          headers:
            Location:
              description: The url to GET the node.
              schema:
                type: string
            X-ID:
              description: The ID of the new file.
              schema:
                type: integer
                format: int64
        '400':
          description: >-
            Returned if the request is not valid and the file could not be
            uploaded.
        '403':
          description: Returned if the user does not have permissions to upload files.
        '404':
          description: Returned if the parentID does not exist.
        '409':
          description: Returned if a node with the given name already exists.
        '412':
          description: Returned if the parentID is a file instead of a directory.
        '413':
          description: Returned if you attempt to upload a file that is too large.
        '422':
          description: Returned if the uploaded file size was incorrect.
        '423':
          description: Returned if the path is locked.
        '451':
          description: Returned if the share is in readonly mode.
        '500':
          description: Returned if the server errored and could not complete the request for unknown reasons.
        '503':
          description: Returned if the share is currently unavailable or was disabled while transferring. Retry the request later.
        '507':
          description: Returned if the share has insufficient storage for the file.
...
servers:
...
components:
  securitySchemes:
    pre2FAToken:
      name: X-Pre2FAToken
      in: header
      type: apiKey
      description: >
        Uses a pre 2FA session token to authenticate. Use `/authenticate` to
        obtain a pre 2FA session token. Only issued when Two Factor
        Authentication is enabled for the user. Can only be used to verify a 2FA
        challenge and to obtain an `X-SessionToken`.
    sessionToken:
      name: X-SessionToken
      in: header
      type: apiKey
      description: >-
        Uses a session token to authenticate. Use `/authenticate` to obtain a
        session token.
    sessionCookie:
      name: SessionToken
      in: cookie
      type: apiKey
      description: >-
        Uses a session token to authenticate. Use `/authenticate` to obtain a session token.
        When using a cookie to authenticate, a secondary CSRF token is required to be present,
        either as `X-CSRF-Token` header or as GET `CSRF-Token` parameter.
        The CSRF token can be acquired directly from `/authenticate` or by GETting `/authenticate/csrf-token`
    sudoToken:
      name: X-SudoToken
      in: header
      type: apiKey
      description: >-
        Uses a sudo token to authenticate. Use `/authenticate/sudo` to obtain a sudo token.
    appToken:
      name: X-AppToken
      in: header
      type: apiKey
      description: >-
        Uses a App token to authenticate. Use `/authenticate/tokens` to obtain an app token.
        To authenticate a request using an App token specify the header `X-AppToken` with the given token string.

        For `GET` requests where custom headers are not possible, it is allowed to specify the token string in a `GET` parameter with the name `App-Token`,
...
  schemas:
...
    authentication:
      type: object
      required:
      - username
      - password
      properties:
        username:
          type: string
          example: s_jansen
        password:
          type: string
          example: SecritPassw0rd
        sessionDuration:
          type: integer
          example: 3600
          default: 86400
          minimum: 60
          maximum: 7776000
          description: The session duration in seconds.
...
    shareEvent:
      type: object
      properties:
        shareId:
          type: integer
          format: int64
          description: The ID of the share.
        path:
          type: string
          description: Path to the node that is shared.
          example: /files/path/to/my/share
        nodeId:
          type: integer
          format: int64
          description: The node id of the node that is shared.
        type:
          type: string
          description: The type of the share.
          enum:
            - Public
            - Internal
        isDir:
          type: boolean
          description: Is the share a directory?
        hasPassword:
          type: boolean
          description: Is there a password on the share?
        expiresAt:
          type: integer
          format: int64
          description: The unix timestamp of the share creation.
        description:
          type: string
          description: The share description.
        urlToken:
          type: string
          description: The URL token used to access the share
        permissions:
          $ref: '#/components/schemas/sharePermissions'
        token:
          type: string
          description: Use URLToken instead.
          deprecated: true
        allowUpload:
          type: boolean
          deprecated: true
          description: Use permissions instead.
    getShareEvent:
      type: object
      properties:
        eventId:
          type: integer
          format: uint64
          description: A unique identifier for this event.
          example: 146295914347823327
        time:
          type: integer
          example: 1462959140
          description: The unix timestamp of the event.
        type:
          type: string
          enum:
            - create
            - update
            - delete
          example: update
        userId:
          type: integer
          example: 1
          description: The user id of the user the auth token is from.
        username:
          type: string
          example: testuser
          description: The username of the user the auth token is from.
        instigator:
          type: string
          example: otheruser
          description: The person/system that caused the event.
        data:
          $ref: '#/components/schemas/shareEvent'
    listShareEvents:
      type: object
      required:
        - events
      properties:
        events:
          type: array
          items:
            $ref: '#/components/schemas/getShareEvent'
...
    nodeMetadata:
      type: object
      required:
      - mediaType
      properties:
        type:
          type: string
          enum:
          - none
          - image
          - video
          - audio
          example: video
          description: The type of media file.
        mediaType:
          type: integer
          enum:
          - 0
          - 1
          - 2
          - 3
          example: 1
          description: 'Integer variant of the type property, 0 = none, 1 = image, 2 = video, 3 = audio'
        width:
          type: integer
          example: 420
        height:
          type: integer
          example: 420
        orientation:
          type: integer
          example: 2
        aspectRatio:
          type: number
          example: 1.3333
        durationMs:
          type: integer
          format: uint32
          example: 458000
...
    sharePermissions:
      type: object
      description: A list of permissions for a share.
      properties:
        createFile:
          type: boolean
        createDirectory:
          type: boolean
        readFile:
          type: boolean
        readDirectory:
          type: boolean
        updateFile:
          type: boolean
        updateDirectory:
          type: boolean
        deleteFile:
          type: boolean
        deleteDirectory:
          type: boolean
...
    shareToken:
      type: object
      properties:
        password:
          type: string
          example: SecritPassw0rd
          description: The password for the share, optional.
    getShareByToken:
      type: object
      properties:
        sid:
          type: integer
          example: 4242
          description: The ID of the share. _Note_, this is separate from the Node ID.
        shareURLToken:
          type: string
          example: dfGkg6545hFv
          description: The token with which the share is accessed.
        shareID:
          type: string
          deprecated: true
          example: dfGkg6545hFv
          description: ! Do not USE ! The URL token with which the share is accessed. Use `shareURLToken` instead.
        allowUpload:
          type: boolean
          example: false
          description: Whether anonymous users are allowed to create directories and upload files inside the share.
        description:
          type: string
          example: Hello, this is a message left by the creator of this share. Download my stuff!
          description: An optional message that can be set on the share.
        language:
          type: string
          example: nl_NL
          description: The language with which the public share should be displayed.
        id:
          type: integer
          format: int64
          example: 4242424242
          description: The ID of the shares root node
        name:
          type: string
          example: ThisIsMyFirstFile.pdf
          description: The name of the shares root node
        dir:
          type: boolean
          example: false
          description: Whether the shares root node is a directory
        mime:
          type: string
          example: image/png
          description: The mime type of the shares root node
        size:
          type: integer
          format: int64
          example: 104200124
          description: The size in bytes of the shares root node
        created:
          type: integer
          format: int64
          example: 34587357
          description: Unix timestamp
        modified:
          type: integer
          format: int64
          example: 34587357
          description: Unix timestamp
        childCount:
          type: integer
          example: 42
          description: Number of direct children nodes
        metadata:
          $ref: '#/components/schemas/nodeMetadata'
...